Removed ORM\Collection#__call
The intention of this method was probably something like chaining conditions.

In the meantime, it was never used and recently caused that an error was not
thrown when it should have been thrown:

$a = Models\Model::where(...);
$a->move();

A collection does not have a move method, yet it failed silently without any
result. What's actually missing there is a call to #first.

$a = Models\Model::where(...)->first();
$a->move();

Now an error is thrown when calling a method on a collection that does not
actually exist. Hooray!